+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## Unreleased
+- Remove temporary dependencies
+- Add Ruby 3.2 to CI
+- Require Rails 6+
+- Require Ruby 2.7+
+
 ## [2.6.4](https://github.com/owen2345/camaleon-cms/tree/2.6.4) (2022-06-08)
 - Reformat JSON comments for OJ compatibility
 - Use OJ in testing